Move controllers spec helper under support
Move controllers spec helper under support
